+++ pushapkscript/script.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 """ PushAPK main script
 """
+import contextlib
 import logging
 import os
 
@@ -18,7 +19,7 @@ async def async_main(context):
     product_config = _get_product_config(context, android_product)
 
     logging.getLogger('oauth2client').setLevel(logging.WARNING)
-    _log_warning_forewords(android_product, context.task['payload'])
+    _log_warning_forewords(not bool(context.config.get('do_not_contact_google_play')), context.task['payload'])
 
     log.info('Verifying upstream artifacts...')
     artifacts_per_task_id, failed_artifacts_per_task_id = artifacts.get_upstream_artifacts_full_paths_per_task_id(context)
@@ -37,17 +38,18 @@ async def async_main(context):
 
     if product_config['update_google_play_strings']:
         log.info('Finding whether Google Play strings can be updated...')
-        google_play_strings_path = googleplay.get_google_play_strings_path(
+        strings_path = googleplay.get_google_play_strings_path(
             artifacts_per_task_id, failed_artifacts_per_task_id
         )
     else:
         log.warning('This product does not upload strings automatically. Skipping Google Play strings search.')
-        google_play_strings_path = None
+        strings_path = None
 
     log.info('Delegating publication to mozapkpublisher...')
-    googleplay.publish_to_googleplay(
-        context.task['payload'], android_product, product_config, all_apks_paths, google_play_strings_path,
-    )
+    with contextlib.ExitStack() as stack:
+        files = [stack.enter_context(open(apk_file_name)) for apk_file_name in all_apks_paths]
+        strings_file = stack.enter_context(open(strings_path)) if strings_path is not None else None
+        googleplay.publish_to_googleplay(context.task['payload'], android_product, product_config, files, strings_file)
 
     log.info('Done!')
 
@@ -65,15 +67,15 @@ def _get_product_config(context, android_product):
     Are you sure you allowed to push such APK?'.format(android_product))
 
 
-def _log_warning_forewords(android_product, task_payload):
-    if googleplay.is_allowed_to_push_to_google_play(android_product):
+def _log_warning_forewords(contact_google_play, task_payload):
+    if contact_google_play:
         if googleplay.should_commit_transaction(task_payload):
             log.warning('You will publish APKs to Google Play. This action is irreversible,\
 if no error is detected either by this script or by Google Play.')
         else:
             log.warning('APKs will be submitted to Google Play, but no change will not be committed.')
     else:
-        log.warning('You do not have the rights to reach Google Play. *All* requests will be mocked.')
+        log.warning('This pushapk instance is not allowed to talk to Google Play. *All* requests will be mocked.')
 
 
 def get_default_config():