+++ pushapkscript/test/test_script.py
@@ -3,10 +3,11 @@ import pytest
 import os
 
 from scriptworker import client, artifacts
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 from pushapkscript import googleplay, jarsigner, task, manifest
 from pushapkscript.script import async_main, get_default_config, main, _log_warning_forewords
+from pushapkscript.test.helpers.mock_file import mock_open
 
 
 @pytest.mark.asyncio
@@ -26,7 +27,6 @@ async def test_async_main(monkeypatch, android_product, update_google_play_strin
             'someOtherTaskId': ['/some/path/to/yet_another.apk', ],
         }, {})
     )
-    monkeypatch.setattr(googleplay, 'is_allowed_to_push_to_google_play', lambda _: False)
     monkeypatch.setattr(jarsigner, 'verify', lambda _, __: None)
     monkeypatch.setattr(manifest, 'verify', lambda _, __: None)
     monkeypatch.setattr(task, 'extract_android_product_from_scopes', lambda _: android_product)
@@ -44,20 +44,24 @@ async def test_async_main(monkeypatch, android_product, update_google_play_strin
     monkeypatch.setattr(googleplay, 'get_google_play_strings_path', google_play_strings_call)
 
     context = MagicMock()
+    context.config = {
+        'do_not_contact_google_play': True
+    }
     context.task = {
         'payload': {}
     }
 
-    def assert_google_play_call(_, __, ___, all_apks_paths, google_play_strings_path):
-        assert sorted(all_apks_paths) == ['/some/path/to/another.apk', '/some/path/to/one.apk', '/some/path/to/yet_another.apk']
+    def assert_google_play_call(_, __, ___, all_apks_files, google_play_strings_file):
+        assert sorted([file.name for file in all_apks_files]) == ['/some/path/to/another.apk', '/some/path/to/one.apk', '/some/path/to/yet_another.apk']
         if android_product == 'focus':
-            assert google_play_strings_path is None
+            assert google_play_strings_file is None
         else:
-            assert google_play_strings_path == '/some/path.json'
+            assert google_play_strings_file.name == '/some/path.json'
 
     monkeypatch.setattr(googleplay, 'publish_to_googleplay', assert_google_play_call)
 
-    await async_main(context)
+    with patch('pushapkscript.script.open', new=mock_open):
+        await async_main(context)
     assert next(google_play_strings_call_counter) == expected_strings_call_count
 
 
@@ -65,13 +69,12 @@ async def test_async_main(monkeypatch, android_product, update_google_play_strin
     (True, True, 'You will publish APKs to Google Play. This action is irreversible,\
 if no error is detected either by this script or by Google Play.'),
     (True, False, 'APKs will be submitted to Google Play, but no change will not be committed.'),
-    (False, False, 'You do not have the rights to reach Google Play. *All* requests will be mocked.'),
-    (False, True, 'You do not have the rights to reach Google Play. *All* requests will be mocked.'),
+    (False, False, 'This pushapk instance is not allowed to talk to Google Play. *All* requests will be mocked.'),
+    (False, True, 'This pushapk instance is not allowed to talk to Google Play. *All* requests will be mocked.'),
 ))
 def test_log_warning_forewords(caplog,  monkeypatch, is_allowed_to_push, should_commit_transaction, expected):
-    monkeypatch.setattr(googleplay, 'is_allowed_to_push_to_google_play', lambda _: is_allowed_to_push)
     monkeypatch.setattr(googleplay, 'should_commit_transaction', lambda _: should_commit_transaction)
-    _log_warning_forewords('product', MagicMock())
+    _log_warning_forewords(not is_allowed_to_push, MagicMock())
     assert len(caplog.records) == 1
     assert caplog.records[0].levelname == 'WARNING'
     assert expected in caplog.text